baisoft-marketplace/requirements.txt
setuptools>=68.0.0
Django==6.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
gunicorn==21.2.0
whitenoise==6.6.0
python-decouple==3.8
psycopg2-binary==2.9.9
Pillow==10.2.0
